🚚 [Move]: #84 number-ex/string-ex を ext/ 配下へ、option/result を utils/ 配下へ再配置
関心ごとにグルーピングするため、`packages/core/src/` の以下を移動:
- `number-ex/` → `ext/number/`（テスト名も -ex 除去）
- `string-ex/` → `ext/string/`（同上）
- `option/`   → `utils/option/`
- `result/`   → `utils/result/`

併せて全 import パスとルート `src/index.ts` の re-export パスを更新。
エクスポートシンボル名 (`NumberEx`, `StringEx`, `Option`, `Result`) と
外部 API は不変。
