+++ packages/core/src/objects/lru-cache/index.ts
@@ -1,5 +1,5 @@
-import type { Result } from "../../result/index";
-import { err, ok } from "../../result/index";
+import type { Result } from "../../utils/result/index";
+import { err, ok } from "../../utils/result/index";
 
 const DefaultCacheCapacity = 1024;
 const MinCacheCapacity = 1;
