+++ packages/core/src/objects/object-parser/direct-object/index.ts
@@ -1,8 +1,8 @@
 import type { PdfParseError } from "../../../errors/index";
-import type { Option } from "../../../option/index";
-import { none, some } from "../../../option/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Option } from "../../../utils/option/index";
+import { none, some } from "../../../utils/option/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { GenerationNumber } from "../../../types/generation-number/index";
 import type { Token } from "../../../types/index";
