+++ packages/core/src/objects/object-parser/indirect-object/index.ts
@@ -1,7 +1,7 @@
 import type { PdfParseError } from "../../../errors/index";
 import { Tokenizer } from "../../../lexer/tokenizer/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { GenerationNumber } from "../../../types/generation-number/index";
 import { TokenType } from "../../../types/index";
