+++ packages/core/src/objects/object-parser/object-parser.indirect.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "vitest";
 import type { PdfError } from "../../errors/index";
-import type { Result } from "../../result/index";
-import { ok } from "../../result/index";
+import type { Result } from "../../utils/result/index";
+import { ok } from "../../utils/result/index";
 import { ByteOffset } from "../../types/byte-offset/index";
 import type { PdfDictionary, PdfObject } from "../../types/pdf-types/index";
 import type { ObjectResolver } from "./index";
