+++ packages/core/src/objects/object-parser/stream-object/index.ts
@@ -1,7 +1,7 @@
 import type { PdfError, PdfParseError } from "../../../errors/index";
 import { isPdfTokenBoundary } from "../../../lexer/bytes/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { GenerationNumber } from "../../../types/generation-number/index";
 import { ObjectNumber } from "../../../types/object-number/index";
