+++ packages/core/src/objects/object-parser/stream-object/stream-object.resolve-length.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test } from "vitest";
-import { err, ok } from "../../../result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import type { ObjectResolver, StreamLength } from "../types";
 import { StreamObject } from "./index";
