+++ packages/core/src/objects/object-parser/types.ts
@@ -1,5 +1,5 @@
 import type { PdfError } from "../../errors/index";
-import type { Result } from "../../result/index";
+import type { Result } from "../../utils/result/index";
 import type { ByteOffset } from "../../types/byte-offset/index";
 import type { GenerationNumber } from "../../types/generation-number/index";
 import type { ObjectNumber } from "../../types/object-number/index";
