+++ packages/core/src/objects/object-store/entry-readers/inline.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test, vi } from "vitest";
-import { err, ok } from "../../../result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { GenerationNumber } from "../../../types/generation-number/index";
 import type { XRefUsedEntry } from "../../../types/pdf-types/index";
