+++ packages/core/src/objects/object-store/object-store.test.helpers.ts
@@ -1,5 +1,5 @@
 import { expect } from "vitest";
-import type { Result } from "../../result/index";
+import type { Result } from "../../utils/result/index";
 import { GenerationNumber } from "../../types/generation-number/index";
 import { ObjectNumber } from "../../types/object-number/index";
 import type {
