+++ packages/core/src/objects/object-stream-extractor/body/index.ts
@@ -1,6 +1,6 @@
 import type { PdfError } from "../../../errors/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import type { ObjectNumber } from "../../../types/object-number/index";
 import type { PdfValue } from "../../../types/pdf-types/index";
