+++ packages/core/src/objects/object-stream-extractor/body/object-stream-body.extract.test.ts
@@ -1,5 +1,5 @@
 import { assert, expect, test, vi } from "vitest";
-import { err, ok } from "../../../result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { GenerationNumber } from "../../../types/generation-number/index";
 import { ObjectNumber } from "../../../types/object-number/index";
