+++ packages/core/src/objects/object-stream-extractor/body/object-stream-body.test.helpers.ts
@@ -1,5 +1,5 @@
 import type { PdfError } from "../../../errors/index";
-import type { Result } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
 import type {
   PdfDictionary,
   PdfObject,
