+++ packages/core/src/objects/object-stream-extractor/flate-decompressor/index.ts
@@ -1,5 +1,5 @@
 import type { PdfParseError } from "../../../errors/index";
-import type { Result } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
 import { decompressFlate } from "../../../xref/stream/flatedecode/index";
 import type { CreateFlateDecompressorOptions } from "../types";
