+++ packages/core/src/objects/object-stream-extractor/header/index.ts
@@ -1,8 +1,8 @@
 import type { PdfParseError } from "../../../errors/index";
 import { Tokenizer } from "../../../lexer/index";
-import { NumberEx } from "../../../number-ex/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import { NumberEx } from "../../../ext/number/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import { ByteOffset } from "../../../types/byte-offset/index";
 import { ObjectNumber } from "../../../types/object-number/index";
 import { TokenType } from "../../../types/pdf-types/index";
