+++ packages/core/src/objects/object-stream-extractor/types.ts
@@ -1,5 +1,5 @@
 import type { PdfError } from "../../errors/index";
-import type { Result } from "../../result/index";
+import type { Result } from "../../utils/result/index";
 import type { ObjectNumber } from "../../types/object-number/index";
 import type { PdfObject } from "../../types/pdf-types/index";
