+++ packages/core/src/types/generation-number/index.ts
@@ -1,5 +1,5 @@
-import type { Result } from "../../result/index";
-import { err, ok } from "../../result/index";
+import type { Result } from "../../utils/result/index";
+import { err, ok } from "../../utils/result/index";
 import type { Brand } from "../brand/index";
 
 declare const GenerationNumberBrand: unique symbol;
