+++ packages/core/src/utils/result/to-option/index.ts
