+++ packages/core/src/xref/merger/index.ts
@@ -1,7 +1,7 @@
 import type { PdfParseError } from "../../errors/index";
-import { NumberEx } from "../../number-ex/index";
-import type { Err, Result } from "../../result/index";
-import { err, ok } from "../../result/index";
+import { NumberEx } from "../../ext/number/index";
+import type { Err, Result } from "../../utils/result/index";
+import { err, ok } from "../../utils/result/index";
 import type { ByteOffset } from "../../types/byte-offset/index";
 import type { TrailerDict, XRefEntry, XRefTable } from "../../types/index";
 import type { ObjectNumber } from "../../types/object-number/index";
