+++ packages/core/src/xref/merger/merger.error.test.ts
@@ -1,7 +1,7 @@
 import { assert, expect, test } from "vitest";
 import type { PdfParseError } from "../../errors/index";
-import type { Result } from "../../result/index";
-import { err, ok } from "../../result/index";
+import type { Result } from "../../utils/result/index";
+import { err, ok } from "../../utils/result/index";
 import { ByteOffset } from "../../types/byte-offset/index";
 import { GenerationNumber } from "../../types/generation-number/index";
 import type { TrailerDict, XRefEntry, XRefTable } from "../../types/index";
