+++ packages/core/src/xref/startxref/scanner/index.ts
@@ -4,8 +4,8 @@ import {
   matchesBytesAt,
   skipWhitespaceAndComments,
 } from "../../../lexer/bytes/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import {
   type ByteOffset,
   ByteOffset as ByteOffsetCompanion,
