+++ packages/core/src/xref/table/parser/index.ts
@@ -4,10 +4,10 @@ import {
   matchesBytesAt,
   skipWhitespaceAndComments,
 } from "../../../lexer/bytes/index";
-import type { Option } from "../../../option/index";
-import { none, some } from "../../../option/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Option } from "../../../utils/option/index";
+import { none, some } from "../../../utils/option/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import {
   type ByteOffset,
   ByteOffset as ByteOffsetCompanion,
