+++ packages/core/src/xref/trailer/parser/index.ts
@@ -5,8 +5,8 @@ import {
   skipWhitespaceAndComments,
 } from "../../../lexer/bytes/index";
 import { Tokenizer } from "../../../lexer/tokenizer/index";
-import type { Result } from "../../../result/index";
-import { err, ok } from "../../../result/index";
+import type { Result } from "../../../utils/result/index";
+import { err, ok } from "../../../utils/result/index";
 import {
   ByteOffset as BO,
   type ByteOffset,
